requirements.txt
streamlit == 1.28.0
pandas == 2.0.3
nump == 1.24.3
plotly == 5.15.0
scikit-learn == 1.3.0
scipy == 1.10.1